refactor(routing): consolidate naming and helper location (U-review-2)
- Finding #16: rename RouteIntentInputs booleans to is* prefix.
  `practicesLoading` → `isPracticesLoading`, `subscriptionSyncInFlight` →
  `isSubscriptionSyncInFlight` on the interface, the computeRouteIntent
  destructuring, the useAuthRouteIntent useMemo, and the test fixtures.
  `hasPracticeMembership` keeps its has* convention. Upstream resolver
  fields (`useWorkspaceResolver.practicesLoading`) keep their existing
  names; only the intent-input boundary renames.
- Finding #11: move getActiveOrganizationPointer from routeIntent.ts to
  authClient.ts so it lives next to unwrapSessionData and the rest of the
  session normalization. Update all five callers
  (useAuthRouteIntent, useEnsureActiveOrganization, usePracticeManagement,
  SessionContext, index.tsx + the unit test) to import from authClient.
  Update the convention doc to point the canonical-reader pin at the new
  location.
