+++ src/index.tsx
@@ -18,7 +18,8 @@ import { useAuthRouteIntent } from '@/shared/hooks/useAuthRouteIntent';
 import { usePostAuthBounce } from '@/shared/hooks/usePostAuthBounce';
 import { AuthenticatedRouter } from '@/shared/auth/AuthenticatedRouter';
 import { Redirect } from '@/shared/auth/Redirect';
-import { assertNeverIntent, getActiveOrganizationPointer } from '@/shared/auth/routeIntent';
+import { assertNeverIntent } from '@/shared/auth/routeIntent';
+import { getActiveOrganizationPointer } from '@/shared/lib/authClient';
 import { AlertTriangle } from 'lucide-preact';
 import { AppGuard } from '@/app/AppGuard';
 import { AuthBootGate } from '@/app/AuthBootGate';
